Added BLL, excel version
I threw in an extra use case. The user will be able to query data set to
get total contributions & total votes per candidate per year. From this
we will use a Least Square Liner Regression model to calculate the most
likely amount of votes that a give contribution total should yield of
vice vera. We don't need to chart it as it is a user query that requires
an answer not a diagram. However a chart for illustrative purposes could
be good. Explanation & chart in draft chart.xls, 2006 tab.